+++ loki/handler.go
@@ -13,6 +13,7 @@ func NewPluginWithService(service Service) *pluginbinding.Plugin {
 	return pluginbinding.Define(manifestSpec(),
 		pluginbinding.RegisterOperation(testSpec(), service.Test),
 		pluginbinding.RegisterOperation(querySpec(), service.Query),
+		pluginbinding.RegisterOperation(metricSpec(), service.Metric),
 		pluginbinding.RegisterOperation(labelsSpec(), service.Labels),
 		pluginbinding.RegisterOperation(recentLogsSpec(), service.RecentLogs),
 		pluginbinding.RegisterDatasourceSearch(logEntriesDatasourceSpec(), service.LogEntriesDatasource),